Create CSVDownloader2.0.java
Instead of downloading all of the files then combining them, this version combines them as they are read. Takes about the same time if not a little less. Also handles a few errors: User Input Error and Connection Error. Still need to determine how client will use this project in the future and not just for http://cfe.cboe.com/data/historicaldata.aspx, and adjust the program for those needs.